% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrIndex.R
\name{CorrIndex}
\alias{CorrIndex}
\title{CorrIndex}
\usage{
CorrIndex(cross_correlation_matrix)
}
\arguments{
\item{cross_correlation_matrix}{Cross-correlation matrix}
}
\value{
CorrIndex, which means the closeness between S and S_true , is returned.
}
\description{
Calculate the CorrIndex of the cross-correlation matrix of S_true and estimated S.
The closer the value is to 0, the closer estimated S is to S_true.
}
\examples{
S_true <- matrix(runif(5*5), nrow=5, ncol=5)
S <- matrix(runif(5*5), nrow=5, ncol=5)
CorrIndex(cor(S_true, S))
}
